refactor(api) update seeds for pixPlusEDU Version

# The commit message #2 will be skipped:

# fixup! refactor(api) update seeds for pixPlusEDU Version
